Fix Bundling. Fixes #50 Fixes #39 Fixes #10.
This module uses Infinite Ajax Scroll v2.2.1. All the plugins for
that module (Pagination etc) are not compatible with requireJs because
they are using globals. The only way that this modules JS can access
these globals is by either converting IAS to a requireJs compatible
library or by including the entire file in the infinitescroll.js. This
is a hack, I am aware of that, but until IAS 3.x is available, I think
this at least returns the possiblility to bundle, merge and minify js
again in Magento stores using this plugin.